move lazy to its own file due to circular dependencies introduced
