+++ thoth/python/artifact.py
@@ -29,7 +29,7 @@ from elftools.elf.elffile import ELFFile
 from typing import Iterator, Tuple
 import attr
 
-from .helpers import Lazy
+from .lazy import Lazy
 
 _LOGGER = logging.getLogger(__name__)
